plot correlations

+++ src/zuy/semeds/save_outputs.py
@@ -6,11 +6,15 @@ import pandas as pd  # type: ignore
 
 from zuy.common.dftools import save_formatted_xlsx
 from zuy.semeds.models import ZakazkaSample
+from zuy.semeds.plot_element_correlations import plot_correlations_from_tsv
+
 
 logger = logging.getLogger(__name__)
 
 
-def save_output(samples: dict[ZakazkaSample, pd.DataFrame], outdir: Path) -> None:
+def save_output(
+    samples: dict[ZakazkaSample, pd.DataFrame], outdir: Path, overwrite: bool = False
+) -> None:
     """
     Save each sample's DataFrame to LaTeX, TSV, XLS, and PNG heatmap.
     """
@@ -23,21 +27,28 @@ def save_output(samples: dict[ZakazkaSample, pd.DataFrame], outdir: Path) -> Non
 
         # 1. Save .xls
         fp_xls = outdir / f"{zak_sample}.xls"
-        save_formatted_xlsx(fp_xls, {"results": df}, widths=[(0, 50, 5)])
+        if fp_xls.exists() and not overwrite:
+            logger.info(
+                f"File {fp_xls} exists and overwrite is False. Skipping XLS, TEX, TSV save."
+            )
+        else:
+            save_formatted_xlsx(fp_xls, {"results": df}, widths=[(0, 50, 5)])
 
-        df2 = df.droplevel(["Zakazka", "Sample"]).drop(
-            columns=["source_file", "Site", "Project"]
-        )
+            df2 = df.droplevel(["Zakazka", "Sample"]).drop(
+                columns=["source_file", "Site", "Project"]
+            )
 
-        # drop unneeded index levels
-        # 2. Save .tex
-        tex = df2.to_latex(index=True, na_rep="", float_format="%.1f")
-        tex = re.sub(r" &", "\t&", tex)  # tab-align
-        fp_tex = outdir / f"{zak_sample}.tex"
-        fp_tex.write_text(tex)
+            # drop unneeded index levels
+            # 2. Save .tex
+            fp_tex = outdir / f"{zak_sample}.tex"
 
-        # 3. Save .tsv
-        fp_tsv = outdir / f"{zak_sample}.tsv"
-        df2.to_csv(fp_tsv, sep="\t", index=True, na_rep="", float_format="%.1f")
+            tex = df2.to_latex(index=True, na_rep="", float_format="%.1f")
+            tex = re.sub(r" &", "\t&", tex)  # tab-align
+            fp_tex.write_text(tex)
+
+            # 3. Save .tsv
+            fp_tsv = outdir / f"{zak_sample}.tsv"
+            df2.to_csv(fp_tsv, sep="\t", index=True, na_rep="", float_format="%.1f")
+            plot_correlations_from_tsv(fp_tsv)
 
     logger.info(f"Saved outputs to {outdir}")
